domparsing/Range-createContextualFragment.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<html>
<head>
  <title>domparsing Test: range.createContextualFragment</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
 </head>
<body>
   <h1>domparsing-range-createContextualFragment</h1>
<script>
function docFragment(markupString) {
  var range = document.createRange();
  var frag = range.createContextualFragment(markupString);
  return frag;
}
test(function() {
	var docFrag = docFragment('<foo></foo>');
	assert_equals(docFrag.baseURI, document.URL);
	assert_equals(docFrag.nodeName, "#document-fragment");
	assert_equals(docFrag.nodeType, document.DOCUMENT_FRAGMENT_NODE);
	assert_equals(docFrag.nodeValue, null);
}, 'check creatContextualFragment return a DocumentFragment');
test(function() {
	var docFrag = docFragment('<foo></foo>');
	assert_equals(docFrag.ownerDocument, document);
}, 'check createContextualFragment create an HTML document');
test(function() {
  var markupStr = '<foo1>test1</foo1><foo2>test2</foo2>';
	var docFrag = docFragment(markupStr);
	assert_equals(docFrag.childNodes.length, "2");
	assert_equals(docFrag.textContent, "test1test2");
}, 'check createContextualFragment include all markup tree nodes');
</script>
</body>
</html>
